# Optimize `CliffWalkingFunctional` `reward`