@amacati 2025-09-03
See https://github.com/Farama-Foundation/Gymnasium/pull/1433#issuecomment-3248182204.